created a new component for NoGeoLocation error modal
